requirements.txt
# Core dependencies for PDF table extraction
pandas>=1.3.0
tabula-py>=2.5.0

# Optional dependencies for advanced table detection
# Uncomment the following lines if you want to use camelot-py
# camelot-py[cv]>=0.10.0
# opencv-python>=4.5.0

# Additional utilities
pathlib2>=2.3.0; python_version < "3.4"